# apis/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "薄云商机倍增服务 API",
    "version": "1.0.0",
    "description": "薄云商机倍增服务 Enterprise Management Platform API"
  },
  "servers": [],
  "security": [],
  "paths": {
...
    "/api-docs.json": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api-docs.json"
      }
    },
...
    "/api/v1/auth/companies": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [
          "认证管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "获取当前用户可访问的公司列表",
        "summary": "GET /api/v1/auth/companies"
      }
    },
...
    "/api/v1/auth/projects": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [
          "认证管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "获取指定公司下可访问的项目列表",
        "summary": "GET /api/v1/auth/projects"
      }
    },
...
    "/api/v1/companies/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "公司管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/companies/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "公司管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/companies/"
      }
    },
...
    "/api/v1/skills/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "技能管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/skills/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "技能管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/skills/"
      }
    },
...
    "/api/v1/users/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/users/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/users/"
      }
    },
...
    "/api/v1/llm-models/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "LLM 模型"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/llm-models/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "LLM 模型"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/llm-models/"
      }
    },
...
    "/api/v1/system-configs/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "系统配置"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/system-configs/"
      },
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "系统配置"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /api/v1/system-configs/"
      }
    },
...
    "/api/v1/publishing-platforms/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [
          "发布平台"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/publishing-platforms/"
      }
    },
    "/api/v1/projects/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "项目管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/projects/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "项目管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/projects/"
      }
    },
...
    "/api/v1/upload/": {
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "文件上传"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/upload/"
      }
    },
...
    "/api/v1/publishing-schedule/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/__object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "tags": [
          "发布计划"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/publishing-schedule/"
      }
    },
...
    "/api/v1/todos/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "待办管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/v1/todos/"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [],
        "tags": [
          "待办管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /api/v1/todos/"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "__object": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      }
    }
  },
...
}